Added a leaderboard

+++ src/commands/counting.ts
@@ -1,33 +1,42 @@
-import { ChatInputCommandInteraction, PermissionFlagsBits, SlashCommandBuilder } from 'discord.js';
+import { ChatInputCommandInteraction, GuildChannel, MessageFlags, PermissionFlagsBits, SlashCommandBuilder, TextChannel } from 'discord.js';
 import database from '../utils/database.js';
 
 const data = database('counting');
 
 export const command = new SlashCommandBuilder()
     .setName('counting')
-    .setDescription('Manages counting')
+    .setDescription('Manages counting.')
+    .addSubcommand(leaderboard => leaderboard
+      .setName('leaderboard')
+      .setDescription('Sends what server has the highest count.')
+    )
     .addSubcommand(set => set
         .setName('set')
-        .setDescription('Sets the current channel for counting')
+        .setDescription('Sets the current channel for counting.')
     )
     .addSubcommand(remove => remove
         .setName('remove')
         .setDescription('Removes the current channel form counting')
         .addBooleanOption(option => option
             .setName('all')
-            .setDescription('Removes every channel from counting')
+            .setDescription('Removes every channel from counting.')
         )
     )
-    .setDefaultMemberPermissions(PermissionFlagsBits.ManageChannels)
 
 export async function execute(interaction: ChatInputCommandInteraction) {
   switch (interaction.options.getSubcommand()) {
+    case 'leaderboard': return leaderboard(interaction);
     case 'set': return set(interaction);
     case 'remove': return remove(interaction);
   }
 }
 
 async function set(interaction: ChatInputCommandInteraction) {
+  if (!interaction.memberPermissions.has(PermissionFlagsBits.ManageChannels)) return interaction.reply({
+    content: '❌ You do not have permission to manage channels.',
+    flags: MessageFlags.Ephemeral
+  });
+  
   const result = await data.updateOne({channel: interaction.channel.id}, {
     $setOnInsert: {
       channel: interaction.channel.id,
@@ -47,6 +56,11 @@ async function set(interaction: ChatInputCommandInteraction) {
 }
 
 async function remove(interaction: ChatInputCommandInteraction) {
+  if (!interaction.memberPermissions.has(PermissionFlagsBits.ManageChannels)) return interaction.reply({
+    content: '❌ You do not have permission to manage channels.',
+    flags: MessageFlags.Ephemeral
+  });
+  
   const result = await data.deleteOne({channel: interaction.channel.id})
 
   if (result.deletedCount > 0) {
@@ -56,4 +70,16 @@ async function remove(interaction: ChatInputCommandInteraction) {
   else {
     await interaction.reply('❌ This channel is not a counting channel.');
   }
+}
+
+async function leaderboard(interaction: ChatInputCommandInteraction) {
+  const document: any = await data.findOne({}, { sort: { count: -1 } });
+
+  if (!document) return interaction.reply(`❌ No counting channels exist.`);
+
+  const guild = (await interaction.client.channels.fetch(document.channel) as GuildChannel)
+    ?.guild;
+
+  if (guild)
+  return interaction.reply(`🏆 **${guild}** currently has the highest count at **${document.count}**!`);
 }
